chore: clippy + doc audit pass
- math: suppress approx_constant/large_const_arrays on generated LUT;
  allow unnecessary_cast (needed in 64-bit Inner modes); allow
  suspicious_{op_assign,arithmetic}_impl (correct fixed-point scaling);
  allow should_implement_trait on shr (ergonomic inherent method kept)
- gameplay: add mut_from_ref allows on intrusive-list traversal methods;
  add repr(u8) to DoorKind/FloorKind/CeilKind; explicit transmute types
  in save.rs; drop redundant *mut T as *mut T no-op casts
- gamestate: drop redundant unsafe casts to load/save_game_from_bytes
- sound/rodio: replace index loops with position()+single-write pattern
- pic-data: collapse nested if-let chains; allow too_many_arguments on
  private build_columns helper; fix empty-line module doc comments
- rbsp: use &mut [T] instead of &mut Vec<T> in compact_vertices; allow
  unnecessary_cast (needed when Float=f32)
- wad/voxel: convert misplaced /// to //! module-level doc comments
- voxel-viewer: explicit lifetime on transmute::<Pixels<'_>, Pixels<'static>>

+++ render/software25d/src/bsp.rs
@@ -98,7 +98,7 @@ impl Software25D {
         rend: &mut impl DrawBuffer,
     ) {
         // TODO: pull duplicate functionality out to a function
-        self.clear(FixedT::from(rend.size().width() as i32));
+        self.clear(FixedT::from(rend.size().width()));
         let mut count = 0;
         // TODO: netupdate
